pyproject.toml
[project]
name = "settleup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crispy-bootstrap5>=2026.3",
    "django>=5.0",
    "django-crispy-forms>=2.6",
    "django-environ>=0.13.0",
    "django-extensions>=4.1",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "whitenoise>=6.12.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.2.0",
    "factory-boy>=3.3.3",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
python_files = ["tests.py", "test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
